fix test, move InviteBanner, delete People
