--- src/screens/People/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "People.js",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./People.js"
-}
